Reorganize the documentation into multiple chapters
The content is still the same. This change is focusing on moving it
around to split the doc into multiple pages.
The order of the toctree is also meant to put the emphasis on the
Session and Element rather than on the driver like previously, because
the driver is the internal API of Mink and beginners should never need
to deal with it except for the instantiation.